broke up to digestable chunks
