+++ app.py
@@ -1,154 +1,84 @@
-import argparse
 import os
-import random
-
-import numpy as np
-import torch
-import torch.backends.cudnn as cudnn
+import csv
 import gradio as gr
+from PIL import Image
+import json
 
 from minigpt4.common.config import Config
-from minigpt4.common.dist_utils import get_rank
 from minigpt4.common.registry import registry
 from minigpt4.conversation.conversation import Chat, CONV_VISION
 
-# imports modules for registration
+# Imports modules for registration
 from minigpt4.datasets.builders import *
 from minigpt4.models import *
 from minigpt4.processors import *
 from minigpt4.runners import *
 from minigpt4.tasks import *
 
-def parse_args():
-    parser = argparse.ArgumentParser(description="Demo")
-    parser.add_argument("--cfg-path", type=str, default='eval_configs/minigpt4.yaml', help="path to configuration file.")
-    parser.add_argument(
-        "--options",
-        nargs="+",
-        help="override some settings in the used config, the key-value pair "
-        "in xxx=yyy format will be merged into config file (deprecate), "
-        "change to --cfg-options instead.",
-    )
-    args = parser.parse_args()
-    return args
+# Importing functions from the new modules
+from config_parser import parse_args
+from utils import setup_seeds, preprocess_image
+from error_handler import handle_error
+from gradio_interface import gradio_reset, upload_img, gradio_ask, gradio_answer
+from model_setup import initialize_model
 
-
-def setup_seeds(config):
-    seed = config.run_cfg.seed + get_rank()
-
-    random.seed(seed)
-    np.random.seed(seed)
-    torch.manual_seed(seed)
-
-    cudnn.benchmark = False
-    cudnn.deterministic = True
-    
-# ========================================
-#             Model Initialization
-# ========================================
-
-SHARED_UI_WARNING = f'''### [NOTE] It is possible that you are waiting in a lengthy queue.
-
-You can duplicate and use it with a paid private GPU.
-
-<a class="duplicate-button" style="display:inline-block" target="_blank" href="https://huggingface.co/spaces/Vision-CAIR/minigpt4?duplicate=true"><img style="margin-top:0;margin-bottom:0" src="https://huggingface.co/datasets/huggingface/badges/raw/main/duplicate-this-space-xl-dark.svg" alt="Duplicate Space"></a>
-
-Alternatively, you can also use the demo on our [project page](https://minigpt-4.github.io).
+SHARED_UI_WARNING = f'''### [NOTE] This is still in beta, problems may occur.
 '''
 
-print('Initializing Chat')
-cfg = Config(parse_args())
-
-model_config = cfg.model_cfg
-model_cls = registry.get_model_class(model_config.arch)
-model = model_cls.from_config(model_config).to('cuda:0')
-
-vis_processor_cfg = cfg.datasets_cfg.cc_align.vis_processor.train
-vis_processor = registry.get_processor_class(vis_processor_cfg.name).from_config(vis_processor_cfg)
-chat = Chat(model, vis_processor)
-print('Initialization Finished')
-
-# ========================================
-#             Gradio Setting
-# ========================================
-
-def gradio_reset(chat_state, img_list):
-    if chat_state is not None:
-        chat_state.messages = []
-    if img_list is not None:
-        img_list = []
-    return None, gr.update(value=None, interactive=True), gr.update(placeholder='Please upload your image first', interactive=False), gr.update(value="Upload & Start Chat", interactive=True), chat_state, img_list
-
-def upload_img(gr_img, text_input, chat_state):
-    if gr_img is None:
-        return None, None, gr.update(interactive=True), chat_state, None
-    chat_state = CONV_VISION.copy()
-    img_list = []
-    llm_message = chat.upload_img(gr_img, chat_state, img_list)
-    return gr.update(interactive=False), gr.update(interactive=True, placeholder='Type and press Enter'), gr.update(value="Start Chatting", interactive=False), chat_state, img_list
-
-def gradio_ask(user_message, chatbot, chat_state):
-    if len(user_message) == 0:
-        return gr.update(interactive=True, placeholder='Input should not be empty!'), chatbot, chat_state
-    chat.ask(user_message, chat_state)
-    chatbot = chatbot + [[user_message, None]]
-    return '', chatbot, chat_state
-
-
-def gradio_answer(chatbot, chat_state, img_list, num_beams, temperature):
-    llm_message = chat.answer(conv=chat_state, img_list=img_list, max_new_tokens=300, num_beams=1, temperature=temperature, max_length=2000)[0]
-    chatbot[-1][1] = llm_message
-    return chatbot, chat_state, img_list
-
 title = """<h1 align="center">Demo of MiniGPT-4</h1>"""
-description = """<h3>This is the demo of MiniGPT-4. Upload your images and start chatting!</h3>"""
-article = """<p><a href='https://minigpt-4.github.io'><img src='https://img.shields.io/badge/Project-Page-Green'></a></p><p><a href='https://github.com/Vision-CAIR/MiniGPT-4'><img src='https://img.shields.io/badge/Github-Code-blue'></a></p><p><a href='https://github.com/TsuTikgiau/blip2-llm/blob/release_prepare/MiniGPT_4.pdf'><img src='https://img.shields.io/badge/Paper-PDF-red'></a></p>
+description = """<h3>This is the demo of MiniGPT-4, brought to you by Wadah Adlan and the team. Upload your images and start chatting!</h3>"""
+article = """<p><a href='https://minigpt-4.github.io'><img src='https://img.shields.io/badge/Project-Page-Green'></a></p><p><a href='https://github.com/Vision-CAIR/MiniGPT-4'><img src='https://img.shields.io/badge/Github-Code-blue'></a></p><p><a href='https://github.com/TsuTikgiau/blip2-llm/blob/release_prepare/MiniGPT_4.pdf'><img src='https://img.shields.io/badge/Paper-PDF-red'></a></p><p><b>Special thanks to Wadah Adlan for his valuable contributions to the MiniGPT-4 project.</b></p>
 """
 
-#TODO show examples below
 
-with gr.Blocks() as demo:
-    gr.Markdown(title)
-    gr.Markdown(SHARED_UI_WARNING)
-    gr.Markdown(description)
-    gr.Markdown(article)
+def main():
+    # Initialize Chat
+    print('Initializing Chat')
+    cfg, model, chat = initialize_model(parse_args())
 
-    with gr.Row():
-        with gr.Column(scale=0.5):
-            image = gr.Image(type="pil")
-            upload_button = gr.Button(value="Upload & Start Chat", interactive=True, variant="primary")
-            clear = gr.Button("Restart")
-            
-            num_beams = gr.Slider(
+    setup_seeds(cfg)
+
+    # Gradio Interface
+    with gr.Interface(
+        fn=gradio_answer,
+        inputs=[
+            gr.inputs.Textbox(
+                label='User', placeholder='Please upload your images first', interactive=False),
+            gr.inputs.Image(type="pil", label="Upload Images", multiple=True),
+            gr.inputs.Slider(
                 minimum=1,
                 maximum=5,
-                value=1,
+                default=1,
                 step=1,
-                interactive=True,
-                label="beam search numbers)",
-            )
-            
-            temperature = gr.Slider(
+                label="Beam Search Numbers",
+                help="The number of beams used during decoding."
+            ),
+            gr.inputs.Slider(
                 minimum=0.1,
                 maximum=2.0,
-                value=1.0,
+                default=1.0,
                 step=0.1,
-                interactive=True,
                 label="Temperature",
-            )
-            
+                help="The temperature value used during decoding."
+            ),
+        ],
+        outputs=[
+            gr.outputs.Textbox(label='Chatbot'),
+            gr.outputs.HTML(label='UI Warning')
+        ],
+        title=title,
+        description=description,
+        article=article,
+        analytics_enabled=False,
+        examples=[
+            ['', '', 1, 1]
+        ],
+        allow_flagging=False,
+        allow_screenshot=False,
+        theme="huggingface",
+    ).launch(inline=False, debug=True):
+        print('Gradio is up...')
 
-        with gr.Column():
-            chat_state = gr.State()
-            img_list = gr.State()
-            chatbot = gr.Chatbot(label='MiniGPT-4')
-            text_input = gr.Textbox(label='User', placeholder='Please upload your image first', interactive=False)
-    
-    upload_button.click(upload_img, [image, text_input, chat_state], [image, text_input, upload_button, chat_state, img_list])
-    
-    text_input.submit(gradio_ask, [text_input, chatbot, chat_state], [text_input, chatbot, chat_state]).then(
-        gradio_answer, [chatbot, chat_state, img_list, num_beams, temperature], [chatbot, chat_state, img_list]
-    )
-    clear.click(gradio_reset, [chat_state, img_list], [chatbot, image, text_input, upload_button, chat_state, img_list], queue=False)
 
-demo.launch(enable_queue=True)
+if __name__ == "__main__":
+    main()
